moved serverutil to settings blueprint
